\name{ST}
\alias{ST}
\docType{data}
\title{Smith-Todd dataset}
\description{
A subset of the Lalonde dataset (see cited reference).
}
\usage{data(ST)}
\format{
  A data frame with 250 observations on the following 10 variables.
  \describe{
    \item{\code{treated}}{treated variable indicator}
    \item{\code{age}}{age}
    \item{\code{education}}{years of education}
    \item{\code{black}}{race indicator variable}
    \item{\code{married}}{marital status indicator variable}
    \item{\code{nodegree}}{indicator variable of not possessing a degree}
    \item{\code{re74}}{real earnings in 1974}
    \item{\code{re75}}{real earnings in 1975}
    \item{\code{re78}}{real earnings in 1978 (post treatment outcome)}
    \item{\code{hispanic}}{ethnic indicator variable}
  }
}
\source{see references}
\references{
Smith, J., Todd, P. (2005a) Does Matching Overcome 
Lalonde's Critique of Nonexperimental Estimators?, \emph{Journal of 
Econometrics}, 125(1-2), 305-353.
}
\keyword{datasets}
